IG-9-1 "code refactoring, замечания устранены"
