+++ src/common/components/DecorationBorder/DecorationBorder.module.scss
@@ -5,7 +5,7 @@
   color: #828384C4;
   font-size: 14px;
 
-  .one, .two {
+  .line {
     width: 108px;
     height: 1px;
     background-color: #828384C4;
