+++ src/common/components/DecorationBorder/DecorationBorder.tsx
@@ -4,9 +4,9 @@ import s from "./DecorationBorder.module.scss";
 export const DecorationBorder: React.FC = () => {
   return (
     <div className={s.decorationBorderWrapper}>
-      <div className={s.one} />
+      <div className={s.line} />
       <div className={s.text}>OR</div>
-      <div className={s.two} />
+      <div className={s.line} />
     </div>
   );
 };
