+++ src/common/components/Footer/ComponentLink/index.ts
@@ -1,0 +1,1 @@
+export { ComponentLink } from "./ComponentLink";
