+++ src/common/components/Footer/Footer.module.scss
@@ -1,41 +1,20 @@
-.wrapperWindow {
-  font-size: 12px;
+.footer {
+  margin-top: 20px;
   height: 150px;
-  color: #8e8e8e;
-}
 
-.barOne {
-  display: flex;
-  padding: 20px 0;
-  flex-direction: column;
-  align-items: center;
-
-  .Link {
+  a, span {
     color: #8e8e8e;
+    font-size: 12px;
   }
-
-  .Link:active {
-    opacity: 0.3;
-  }
-}
-
-.sectionTwo {
-  padding: 10px 0;
 }
 
 .barTwo {
-  justify-content: center;
-
   .nativeSelect {
     font-size: 12px;
     color: #8e8e8e;
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
-
-    select {
-      padding: 0;
-    }
 
     option {
       font-size: 14px;
@@ -50,8 +29,8 @@
   .nativeSelect:after {
     border: none;
   }
-
 }
+
