+++ src/common/components/Header/Header.tsx
@@ -4,7 +4,6 @@ import { SearchHeader } from "./SearchHeader";
 import { NavigationBarHeader } from "./NavigationBarHeader";
 
 export const Header = () => {
-  console.log("Header Load");
   return (
     <header className={s.header}>
       <div className={s.logo}>
