--- src/common/components/Login/LogInForm/LogInForm.module.scss
@@ -1,92 +1,0 @@
-.registrationBarWrapper,
-.forgotYourPasswordWrapper, .loginFbButton {
-  display: flex;
-}
-
-.wrapperLoginForm {
-  label {
-    font-size: 12px;
-  }
-
-  input {
-    font-size: 12px;
-    height: 15px;
-    background-color: #FAFAFA;
-  }
-}
-
-.passwordWrapper {
-  width: 266px;
-}
-
-.sendBtnWrapper {
-  margin-top: 15px;
-}
-
-.decorationBorderWrapper {
-  margin-top: 15px;
-  max-width: 266px;
-}
-
-.loginOnFacebookWrapper {
-  font-size: 14px;
-  margin-top: 20px;
-  width: 266px;
-  height: 32px;
-
-  :active {
-    opacity: 0.3;
-  }
-
-  img {
-    width: 6%;
-  }
-
-  .loginFbButton {
-    justify-content: center;
-    align-items: center;
-    color: #4267b1;
-    gap: 10px;
-    text-decoration: none;
-  }
-
-}
-
-.errorWindow {
-  display: none;
-}
-
-.forgotYourPasswordWrapper {
-  padding: 20px 0;
-  width: 266px;
-  height: 15px;
-  align-items: center;
-  justify-content: center;
-
-  .forgotPasswordBtn {
-    color: #2f4e8f;
-    font-size: 11px;
-    text-decoration: none;
-  }
-
-  .forgotPasswordBtn:active {
-    opacity: 0.3;
-  }
-
-}
-
-.sendBtnWrapper {
-  .loadingButton {
-    width: 100%;
-    height: 32px;
-    background-color: #0095f6;
-    text-transform: none;
-    color: white;
-  }
-
-  .loadingButton:disabled {
-    background-color: #0095f6;
-    color: white;
-    opacity: 0.3;
-  }
-}
