--- src/common/components/PasswordReset/PasswordReset.module.scss
@@ -1,18 +1,0 @@
-.passwordResetWrapper {
-  width: 100%;
-  display: flex;
-
-  header {
-    width: 100%;
-    position: fixed;
-    img{
-      margin-left: 40px;
-      width: 50%;
-    }
-  }
-}
-
-.windowContent{
-  width: 390px;
-
-}
