--- src/common/components/PasswordReset/index.ts
@@ -1,1 +1,0 @@
-export { PasswordReset } from "./PasswordReset";
