--- src/common/components/Registration/Registration.module.scss
@@ -1,117 +1,0 @@
-.wrapperBox {
-  margin: 35px auto 15px;
-  border: 1px solid #dbdbdb;
-  background-color: white;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: #8e8e8e;
-  padding: 30px 40px;
-  width: 350px;
-
-  img {
-    width: 68%;
-  }
-
-  input {
-    font-size: 12px;
-    background-color: #fafafa;
-  }
-
-  label {
-    font-size: 12px;
-  }
-
-  .passwordWrapper {
-    width: 268px;
-  }
-
-  .signUpBtn {
-    height: 33px;
-    background-color: #0095f6;
-    font-size: 14px;
-    text-transform: none;
-  }
-
-  .signUpBtn:disabled {
-    background-color: #0095f6;
-    color: white;
-    opacity: 0.3;
-  }
-
-}
-
-.titleSignUp {
-  font-size: 16px;
-  text-align: center;
-}
-
-.buttonSignInFbWrapper {
-  a {
-    background-color: #0095f6;
-    color: white;
-    font-size: 14px;
-    text-transform: none;
-    width: 268px;
-    height: 33px;
-    gap: 6px;
-  }
-
-  a:hover {
-    background-color: #0095f6;
-  }
-
-  a:active {
-    opacity: 0.5;
-  }
-
-  img {
-    width: 10%;
-  }
-}
-
-.decorationBorder {
-  margin-top: 15px;
-}
-
-.signInWrapper {
-  width: 350px;
-  height: 70px;
-  background-color: white;
-  border: 1px solid #dbdbdb;
-  font-size: 14px;
-  gap: 5px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  a {
-    text-decoration: none;
-    color: #0095f6;
-    font-weight: 600;
-    font-size: 14px;
-  }
-
-  a:hover {
-    cursor: pointer;
-  }
-
-  a:active {
-    opacity: 0.3;
-  }
-}
-
-.bottomContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  footer {
-    margin-top: 30px;
-
-    .wrapperWindow {
-      background-color: #0095f6;
-      font-size: 20px;
-    }
-  }
-}
