+++ src/features/Login/LogInForm/CustomForm/index.ts
@@ -1,0 +1,1 @@
+export { CustomForm } from "./CustomForm";
