+++ src/features/Login/WindowAnimatedPhone/InfinitySlider/InfinitySlider.module.scss
