+++ src/features/Login/WindowAnimatedPhone/InfinitySlider/index.ts
