+++ src/features/Login/WindowAnimatedPhone/WindowAnimatedPhone.module.scss
