+++ src/features/Login/WindowAnimatedPhone/WindowAnimatedPhone.tsx
