+++ src/features/Login/WindowLoginForm/WindowLoginForm.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import s from "./WindowLoginForm.module.scss";
-import logo from "../../../../assets/icon/1b47f9d0e595.png";
-import { BarInstallApps } from "../../BarInstallApps";
+import logo from "../../../assets/icon/1b47f9d0e595.png";
+import { BarInstallApps } from "../../../common/components/BarInstallApps";
 import { Link } from "react-router-dom";
 import { LogInForm } from "../LogInForm";
+import { ROUTE_PATHS } from "../../../common/components/App/App";
 
 export const WindowLoginForm: React.FC = () => {
   return (
@@ -18,7 +19,7 @@ export const WindowLoginForm: React.FC = () => {
       </div>
       <div className={s.registrationBarWrapper}>
         <span>Don't have an account yet?</span>
-        <Link to="/signUp">Sign up</Link>
+        <Link to={ROUTE_PATHS.SignUp}>Sign up</Link>
       </div>
       <BarInstallApps />
     </div>
