+++ src/features/Registration/AgreementPolicy/AgreementPolicy.tsx
