+++ src/features/Registration/AgreementPolicy/index.ts
