+++ src/features/Registration/RegistrationCustomForm/index.ts
@@ -1,0 +1,1 @@
+export { RegistrationCustomForm } from "./RegistrationCustomForm";
